requirements.txt
requests
python-socketio[client]
PySide6
